+++ plugins/perfektblue-can/src/perfektblue_can/__init__.py
@@ -1,3 +1,3 @@
 """Optional CAN analysis plugin."""
 
-__version__ = "2.0.0"
+__version__ = "2.1.0"